refactor: split instance registry repository
